The GPU address buffer becomes u64, so bits > 32 stops falling back to CPU
27dabcf8 fixed the u32 truncation by skipping the GPU whenever any neuron
had bits > 32: train wrote an address truncated mod 2^32 while the sparse
eval kernel computed the full u64 one, and the mismatch showed up as
sub-baseline accuracy at b >= 48. Skipping was the safe patch, not the
fix -- it left every adaptive and multistage genome above 32 bits computing
addresses on CPU.

train_address.metal now writes a ulong from wnn_compute_address_u64,
compute_addresses returns Vec<u64>, and both bits guards are deleted.
MAX_GPU_ADDRESSES halves to 128M so the doubled element width leaves the
peak output buffer at the same 1.0 GB.

The new test dispatches 16/32/33/40/50/64-bit widths to Metal and compares
every address against a CPU reference, then asserts the >32 cases actually
exceed u32::MAX -- otherwise it could not tell a correct buffer from a
truncating one. 111/111 worker, 187/187 controller.

+++ src/wnn/ram/strategies/accelerator/metal_train.rs
@@ -200,7 +200,8 @@ impl MetalTrainer
 	///   num_examples: number of training examples
 	///   words_per_example: ceil(total_input_bits / 64)
 	///
-	/// Returns: [total_neurons * num_examples] u32 addresses
+	/// Returns: [total_neurons * num_examples] u64 addresses (u64 since 28/08/2026 —
+/// a u32 buffer truncated bits > 32 and diverged from the u64 sparse eval path)
 	pub fn compute_addresses(
 		&mut self,
 		packed_input: &[u64],
@@ -208,7 +209,7 @@ impl MetalTrainer
 		neuron_meta: &[NeuronTrainMeta],
 		num_examples: usize,
 		words_per_example: usize,
-	) -> Result<Vec<u32>, String>
+	) -> Result<Vec<u64>, String>
 	{
 		let total_neurons = neuron_meta.len();
 
@@ -252,7 +253,7 @@ impl MetalTrainer
 			&self.device,
 			&mut self.cache.address_buffer,
 			output_size,
-			mem::size_of::<u32>(),
+			mem::size_of::<u64>(),
 		);
 
 		// Dispatch GPU kernel
@@ -288,8 +289,8 @@ impl MetalTrainer
 		)?;
 
 		// Read back results
-		let result_ptr = address_buffer.contents() as *const u32;
-		let results: Vec<u32> = unsafe { std::slice::from_raw_parts(result_ptr, output_size).to_vec() };
+		let result_ptr = address_buffer.contents() as *const u64;
+		let results: Vec<u64> = unsafe { std::slice::from_raw_parts(result_ptr, output_size).to_vec() };
 
 		let elapsed_ms = t0.elapsed().as_secs_f64() * 1000.0;
 		eprintln!(
